🚚 move:'모듈 연결 및 변수명 수정'
